requirements.txt
biopython==1.83
matplotlib==3.8.0
pytest==7.4.0